requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
python-dotenv==1.2.1
sqlparse==0.5.4
whitenoise==6.11.0
